refactor: move shared integration contracts to src/shared/integrations.ts
Shared code belongs under shared/; the top-level integrations/ name is
reserved for the later migration of the github/ and gitlab/ folders.

No behavior change: build clean, 2619/2619 tests pass, lint clean.
